requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
chardet==5.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
docx2pdf==0.1.8
fire==0.7.0
fonttools==4.55.0
kombu==5.4.2
lxml==5.3.0
mysqlclient==2.2.6
numpy==2.1.3
opencv-python-headless==4.10.0.84
pdf2docx==0.5.8
pillow==11.0.0
prompt_toolkit==3.0.48
PyMuPDF==1.24.14
pypandoc==1.14
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
redis==5.2.0
reportlab==4.2.5
six==1.16.0
sqlparse==0.5.2
termcolor==2.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
